+++ R/data.R
@@ -316,7 +316,7 @@
 
 #' Rebound terms
 #' 
-#' This is the list of the rebound terms 
+#' This is the list of the rebound terms.
 #' 
 #' @format A string list with `r length(rebound_terms)` entries.
 #' \describe{
@@ -338,6 +338,32 @@
 #' @examples
 #' rebound_terms
 "rebound_terms"
+
+
+#' LaTeX version of rebound terms
+#'
+#' This is the list of the rebound terms in LaTeX form.
+#'
+#' @format A string list with `r length(latex_rebound_terms)` entries.
+#' \describe{
+#' \item{Re_dempl}{Direct emplacement effect rebound, always 0.}
+#' \item{Re_emb}{Indirect embodied energy effect rebound.}
+#' \item{Re_md}{Indirect maintenance and disposal effect rebound.}
+#' \item{Re_empl}{Emplacement effect rebound.}
+#' \item{Re_dsub}{Direct substitution effect rebound.}
+#' \item{Re_isub}{Indirect substitution effect rebound.}
+#' \item{Re_sub}{Substitution effect rebound.}
+#' \item{Re_dinc}{Direct income effect rebound.}
+#' \item{Re_iinc}{Indirect income effect rebound.}
+#' \item{Re_inc}{Income effect rebound.}
+#' \item{Re_prod}{Indirect productivity effect rebound.}
+#' \item{Re_d}{Sum of all direct rebound effects.}
+#' \item{Re_i}{Sum of all indirect rebound effects.}
+#' \item{Re_tot}{Total rebound.}
+#' }
+#' @examples
+#' latex_rebound_terms
+"latex_rebound_terms"
 
 
 #' Graph types