+++ tests/testthat/test-graphs.R
@@ -1,43 +1,49 @@
 
 test_that("rebound_graphs() works as expected", {
   # This is a mess, because all graphs are on the same plot.
-  graphs <- sample_eeu_data_path() %>% 
+  graphs <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs()
   expect_true(!is.null(graphs))
   
   # Try with only one type of graph
-  graphs_energy <- sample_eeu_data_path() %>% 
+  graphs_energy <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(graph_types = "Energy")
   expect_true(!is.null(graphs_energy))
-  expect_equal(graphs_energy$plot_env$.path_data$graph_type %>% unique(), "Energy")
+  expect_equal(graphs_energy$plot_env$.path_data$graph_type %>% unique() %>% as.character(), "Energy")
   
   # Try with only one case, Car Energy
-  graphs_car_energy <- sample_eeu_data_path() %>% 
+  graphs_car_energy <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(cases = "Car", 
                    graph_types = "Energy")
   expect_true(!is.null(graphs_car_energy))
   expect_equal(graphs_car_energy$plot_env$.path_data$Case %>% unique(), "Car")
-  expect_equal(graphs_car_energy$plot_env$.path_data$graph_type %>% unique(), "Energy")
+  expect_equal(graphs_car_energy$plot_env$.path_data$graph_type %>% unique() %>% as.character(), "Energy")
 
   # Try Car Cost
-  graphs_car_cost <- sample_eeu_data_path() %>% 
+  graphs_car_cost <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(cases = "Car", 
                    graph_types = "Cost")
   expect_true(!is.null(graphs_car_cost))
   expect_equal(graphs_car_cost$plot_env$.path_data$Case %>% unique(), "Car")
-  expect_equal(graphs_car_cost$plot_env$.path_data$graph_type %>% unique(), "Cost")
+  expect_equal(graphs_car_cost$plot_env$.path_data$graph_type %>% unique() %>% as.character(), "Cost")
   
   # Try indexed Car Cost
-  graphs_car_cost <- sample_eeu_data_path() %>% 
+  graphs_car_cost <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(indexed = TRUE,
                    cases = "Car", 
                    graph_types = "Cost")
   expect_true(!is.null(graphs_car_cost))
   expect_equal(graphs_car_cost$plot_env$.path_data$Case %>% unique(), "Car")
-  expect_equal(graphs_car_cost$plot_env$.path_data$graph_type %>% unique(), "Cost")
+  expect_equal(graphs_car_cost$plot_env$.path_data$graph_type %>% unique() %>% as.character(), "Cost")
 
   # Eliminate the grids for Car Cost graph.
-  graphs_car_cost_no_grids <- sample_eeu_data_path() %>% 
+  graphs_car_cost_no_grids <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(cases = "Car", 
                    graph_types = "Cost", 
                    grid_types = NULL)
@@ -45,31 +51,35 @@ test_that("rebound_graphs() works as expected", {
   expect_equal(graphs_car_cost_no_grids$plot_env$.grid_data %>% nrow(), 0)
   
   # Try an Energy graph for lamps
-  graphs_lamp_energy <- sample_eeu_data_path() %>% 
+  graphs_lamp_energy <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(cases = "Lamp", 
                    graph_types = "Energy")
   expect_true(!is.null(graphs_lamp_energy))
   expect_equal(graphs_lamp_energy$plot_env$.path_data$Case %>% unique(), "Lamp")
-  expect_equal(graphs_lamp_energy$plot_env$.path_data$graph_type %>% unique(), "Energy")
+  expect_equal(graphs_lamp_energy$plot_env$.path_data$graph_type %>% unique() %>% as.character(), "Energy")
   
   # Try an indexed Energy graph for lamps
-  graphs_indexed_lamp_energy <- sample_eeu_data_path() %>% 
+  graphs_indexed_lamp_energy <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(indexed = TRUE, 
                    cases = "Lamp", 
                    graph_types = "Energy")
   expect_true(!is.null(graphs_indexed_lamp_energy))
   expect_equal(graphs_indexed_lamp_energy$plot_env$.path_data$Case %>% unique(), "Lamp")
-  expect_equal(graphs_indexed_lamp_energy$plot_env$.path_data$graph_type %>% unique(), "Energy")
+  expect_equal(graphs_indexed_lamp_energy$plot_env$.path_data$graph_type %>% unique() %>% as.character(), "Energy")
   
   # Try two cases
-  graphs_two_cases_indexed_energy <- sample_eeu_data_path() %>% 
+  graphs_two_cases_indexed_energy <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(indexed = TRUE,
                    graph_types = "Energy") +
     ggplot2::facet_wrap(facets = "Case")
   expect_true(!is.null(graphs_two_cases_indexed_energy))
   
   # Request both cases
-  graphs_two_cases_indexed_energy_2 <- sample_eeu_data_path() %>% 
+  graphs_two_cases_indexed_energy_2 <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(cases = c("Car", "Lamp"), 
                    indexed = TRUE,
                    graph_types = "Energy") +
@@ -77,7 +87,8 @@ test_that("rebound_graphs() works as expected", {
   expect_true(!is.null(graphs_two_cases_indexed_energy_2))
 
   # Try a preferences graph for lamps
-  graphs_lamp_prefs <- sample_eeu_data_path() %>% 
+  graphs_lamp_prefs <- load_eeu_data() %>% 
+    rebound_analysis() %>% 
     rebound_graphs(cases = "Lamp", 
                    graph_types = "Preferences")
   expect_true(!is.null(graphs_lamp_prefs))